try with return of f32 all the time

+++ scalar/src/lib.rs
@@ -10,10 +10,10 @@ pub fn pro(a: i32, b: i32) -> i32 {
     a*b
 }
 
-pub fn quo(a: f32, b: f32) -> i32 {
-     (a / b) as i32
+pub fn quo(a: f32, b: f32) -> f32 {
+     a / b
 }
 
-pub fn rem(a: f32, b: f32) -> u32 {
-    (a % b) as u32
+pub fn rem(a: f32, b: f32) -> f32 {
+    a % b 
 }
